Character abilities, director shkoli

+++ Assets/Scripts/Entities/Stats/StatAttributes/AttributeMask.cs
@@ -8,8 +8,8 @@ namespace Assets.Scripts.Entities.Stats.StatAttributes
 {
     internal struct AttributeMask
     {
-        public int BaseValue { get; set; }
-        public int BaseMultiplier { get; set; }
+        public float BaseValue { get; set; }
+        public float BaseMultiplier { get; set; }
         public float MaskMultiplier { get; set; }
     }
 }
